better type declaration

+++ src/main.rs
@@ -20,14 +20,14 @@ fn main() {
     println!("Please enter the WIDTH of the grid: ");
     let mut width = String::new();
     io::stdin().read_line(&mut width).expect("Failed to read line");
-    let width: u32 = width.trim().parse().expect("Please type a number!");
+    let width: usize = width.trim().parse().expect("Please type a number!");
 
     println!("Please enter the HEIGHT of the grid: ");
     let mut height = String::new();
     io::stdin().read_line(&mut height).expect("Failed to read line");
-    let height: u32 = height.trim().parse().expect("Please type a number!");
+    let height: usize = height.trim().parse().expect("Please type a number!");
 
-    let mut grid = vec![vec![' '; width as usize]; height as usize];
+    let mut grid = vec![vec![' '; width]; height];
 
     print_grid(&grid);
 }
